improving mm-pallas kernel, improving generation pipeline and AutoModels

+++ src/easydel/modules/gemma/modeling_gemma_flax.py
@@ -1100,7 +1100,7 @@ class FlaxGemmaForCausalLMModule(nn.Module):
 			)
 		else:
 			lm_logits = self.lm_head(hidden_states)
-		lm_logits = lm_logits.astype(jnp.float32)
+		
 
 		if not return_dict:
 			return (lm_logits,) + outputs[1:]
