+++ src/easydel/modules/phi/modelling_phi_flax.py
@@ -851,7 +851,7 @@ class FlaxPhiForCausalLMModule(nn.Module):
 		else:
 			lm_logits = self.lm_head(res.last_hidden_state)
 
-		lm_logits = lm_logits.astype(jnp.float32)
+		
 
 		if not return_dict:
 			return (lm_logits,) + outputs[1:]
