+++ src/easydel/modules/qwen1/modeling_qwen1_flax.py
@@ -1170,7 +1170,7 @@ class FlaxQwen1ForCausalLMModule(nn.Module):
 		else:
 			lm_logits = self.lm_head(hidden_states)
 
-		lm_logits = lm_logits.astype(jnp.float32)
+		
 
 		if not return_dict:
 			return (lm_logits,) + outputs[1:]
